Split data_manager.py into separate modules
