+++ pyqcrbox/pyqcrbox/data_management/__init__.py
@@ -1,1 +1,3 @@
-from .data_file_manager import DummyDataFileManager, QCrBoxDataFileManager
+from .base_data_file_manager import QCrBoxDataFileManager
+from .dummy_data_file_manager import DummyDataFileManager
+from .nats_data_file_manager import NatsDataFileManager
