+++ pyqcrbox/pyqcrbox/data_management/base_data_file_manager.py
@@ -1,0 +1,2 @@
+class QCrBoxDataFileManager:
+    pass
